fix: changed location of newBooking and getExistentBooking methods into lecture-controllers and fixed minor errors
