+++ server/controllers/user-controller.js
@@ -2,6 +2,7 @@
 const knex = require("./../db");
 const jsonwebtoken = require("jsonwebtoken");
 const bcrypt = require("bcrypt");
+const moment = require("moment");
 require("dotenv").config();
 const jwtSecret = process.env.JWT_SECRET;
 
@@ -66,3 +67,5 @@ exports.login = async (req, res) => {
 exports.logout = async (req, res) => {
   res.clearCookie("token").end();
 };
+
+
