+++ server/routes/lecture-route.js
@@ -1,12 +1,14 @@
 const express = require("express");
 
 // Import lecture-controller
-const lecturebookingRoutes = require("./../controllers/lecture-controller");
+const lectureRoutes = require("./../controllers/lecture-controller");
 
 // Create router
 const router = express.Router();
 
-router.get("/newbooking", lecturebookingRoutes.newBooking);
+router.get("/bookable/:userId", lectureRoutes.getBookingLectures);
+router.get("/:lectureId/book", lectureRoutes.newBooking);
+router.get("/previousbooking", lectureRoutes.getExistentBooking);
 
 // Export router
 module.exports = router;
