--- server/routes/lecture_booking-route.js
@@ -1,13 +1,0 @@
-const express = require("express");
-
-// Import lecture-controller
-const lectureRoutes = require("./../controllers/lecture_booking-controller");
-
-// Create router
-const router = express.Router();
-
-router.get("/studentbooking", lectureRoutes.getBookingLectures);
-router.get("/previousbooking", lectureRoutes.getExistentBooking);
-
-// Export router
-module.exports = router;
